feat(errors): implement AppError class and default error mapper for improved error handling
