+++ apps/backend/src/common/utils/fp-utils.spec.ts
@@ -2,13 +2,12 @@ import type { Logger } from "@nestjs/common";
 import { StatusCodes } from "http-status-codes";
 import { z } from "zod";
 
+import { AppError } from "./drizzle-error-utils";
 import {
-  AppError,
   Failure,
   Success,
   assertFailure,
   assertSuccess,
-  defaultRepositoryErrorMapper,
   failure,
   handleFailure,
   isFailure,
@@ -127,65 +126,6 @@ describe("Functional Programming Utilities", () => {
         const result = success("test");
         expect(() => assertFailure(result)).toThrow();
       });
-    });
-  });
-
-  describe("AppError", () => {
-    it("should create AppError with correct properties", () => {
-      const error = new AppError("message", "CODE", 400, { detail: "test" });
-      expect(error.message).toBe("message");
-      expect(error.code).toBe("CODE");
-      expect(error.statusCode).toBe(400);
-      expect(error.details).toEqual({ detail: "test" });
-    });
-
-    it("should create notFound error with correct defaults", () => {
-      const error = AppError.notFound();
-      expect(error.message).toBe("Resource not found");
-      expect(error.code).toBe("NOT_FOUND");
-      expect(error.statusCode).toBe(StatusCodes.NOT_FOUND);
-    });
-
-    it("should create badRequest error with correct defaults", () => {
-      const error = AppError.badRequest();
-      expect(error.message).toBe("Invalid request data");
-      expect(error.code).toBe("BAD_REQUEST");
-      expect(error.statusCode).toBe(StatusCodes.BAD_REQUEST);
-    });
-
-    it("should create forbidden error with correct defaults", () => {
-      const error = AppError.forbidden();
-      expect(error.message).toBe("Access forbidden");
-      expect(error.code).toBe("FORBIDDEN");
-      expect(error.statusCode).toBe(StatusCodes.FORBIDDEN);
-    });
-
-    it("should create unauthorized error with correct defaults", () => {
-      const error = AppError.unauthorized();
-      expect(error.message).toBe("Unauthorized access");
-      expect(error.code).toBe("UNAUTHORIZED");
-      expect(error.statusCode).toBe(StatusCodes.UNAUTHORIZED);
-    });
-
-    it("should create internal error with correct defaults", () => {
-      const error = AppError.internal();
-      expect(error.message).toBe("Internal server error");
-      expect(error.code).toBe("INTERNAL_ERROR");
-      expect(error.statusCode).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
-    });
-
-    it("should create repositoryError with correct defaults", () => {
-      const error = AppError.repositoryError();
-      expect(error.message).toBe("Repository operation failed");
-      expect(error.code).toBe("REPOSITORY_ERROR");
-      expect(error.statusCode).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
-    });
-
-    it("should create validationError with correct defaults", () => {
-      const error = AppError.validationError();
-      expect(error.message).toBe("Validation error");
-      expect(error.code).toBe("VALIDATION_ERROR");
-      expect(error.statusCode).toBe(StatusCodes.BAD_REQUEST);
     });
   });
 
@@ -370,96 +310,6 @@ describe("Functional Programming Utilities", () => {
       expect(result.isFailure()).toBe(true);
       expect(customMapper).toHaveBeenCalledWith(error);
       expect((result as Failure<AppError>).error.message).toBe("Custom error");
-    });
-  });
-
-  describe("defaultRepositoryErrorMapper", () => {
-    it("should return the AppError if one is passed", () => {
-      const originalError = AppError.badRequest("Bad request");
-      const mappedError = defaultRepositoryErrorMapper(originalError);
-
-      expect(mappedError).toBe(originalError);
-    });
-
-    it('should map "not found" errors correctly', () => {
-      const error = new Error("Entity not found");
-      const mappedError = defaultRepositoryErrorMapper(error);
-
-      expect(mappedError.statusCode).toBe(StatusCodes.NOT_FOUND);
-      expect(mappedError.code).toBe("REPOSITORY_NOT_FOUND");
-    });
-
-    it('should map "no rows" errors correctly', () => {
-      const error = new Error("No rows returned");
-      const mappedError = defaultRepositoryErrorMapper(error);
-
-      expect(mappedError.statusCode).toBe(StatusCodes.NOT_FOUND);
-      expect(mappedError.code).toBe("REPOSITORY_NOT_FOUND");
-    });
-
-    it('should map "does not exist" errors correctly', () => {
-      const error = new Error("Record does not exist");
-      const mappedError = defaultRepositoryErrorMapper(error);
-
-      expect(mappedError.statusCode).toBe(StatusCodes.NOT_FOUND);
-      expect(mappedError.code).toBe("REPOSITORY_NOT_FOUND");
-    });
-
-    it('should map "duplicate" errors correctly', () => {
-      const error = new Error("Duplicate entry");
-      const mappedError = defaultRepositoryErrorMapper(error);
-
-      expect(mappedError.statusCode).toBe(StatusCodes.BAD_REQUEST);
-      expect(mappedError.code).toBe("REPOSITORY_DUPLICATE");
-    });
-
-    it('should map "unique constraint" errors correctly', () => {
-      const error = new Error("unique constraint violation");
-      const mappedError = defaultRepositoryErrorMapper(error);
-
-      expect(mappedError.statusCode).toBe(StatusCodes.BAD_REQUEST);
-      expect(mappedError.code).toBe("REPOSITORY_DUPLICATE");
-    });
-
-    it('should map "already exists" errors correctly', () => {
-      const error = new Error("user already exists");
-      const mappedError = defaultRepositoryErrorMapper(error);
-
-      expect(mappedError.statusCode).toBe(StatusCodes.BAD_REQUEST);
-      expect(mappedError.code).toBe("REPOSITORY_DUPLICATE");
-    });
-
-    it('should map "foreign key" errors correctly', () => {
-      const error = new Error("foreign key constraint failed");
-      const mappedError = defaultRepositoryErrorMapper(error);
-
-      expect(mappedError.statusCode).toBe(StatusCodes.BAD_REQUEST);
-      expect(mappedError.code).toBe("REPOSITORY_REFERENCE");
-    });
-
-    it('should map "reference" errors correctly', () => {
-      const error = new Error("reference integrity constraint violation");
-      const mappedError = defaultRepositoryErrorMapper(error);
-
-      expect(mappedError.statusCode).toBe(StatusCodes.BAD_REQUEST);
-      expect(mappedError.code).toBe("REPOSITORY_REFERENCE");
-    });
-
-    it("should map unknown errors to repository errors", () => {
-      const error = new Error("Unknown error");
-      const mappedError = defaultRepositoryErrorMapper(error);
-
-      expect(mappedError.statusCode).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
-      expect(mappedError.code).toBe("REPOSITORY_ERROR");
-    });
-
-    it("should handle non-Error objects", () => {
-      const error = "String error";
-      const mappedError = defaultRepositoryErrorMapper(error);
-
-      expect(mappedError.message).toBe("String error");
-      expect(mappedError.statusCode).toBe(StatusCodes.INTERNAL_SERVER_ERROR);
-      expect(mappedError.code).toBe("REPOSITORY_ERROR");
     });
   });
